Use spring default json serialization in FileEntry
